compute model hash in engine
