+++ modules/nlu/src/backend/engine.ts
@@ -5,7 +5,6 @@ import _ from 'lodash'
 
 import * as CacheManager from './cache-manager'
 import { initializeTools } from './initialize-tools'
-import { computeModelHash } from './model-hash'
 import { Model } from './model-service'
 import { Predict, PredictInput, Predictors, PredictOutput } from './predict-pipeline'
 import SlotTagger from './slots/slot-tagger'
@@ -47,6 +46,23 @@ export default class Engine implements NLUEngine {
   public static async initialize(bp: typeof sdk, version: NLUVersionInfo): Promise<void> {
     this._tools = await initializeTools(bp, version)
     this._version = version
+  }
+
+  // we might want to make this language specific
+  public computeModelHash(
+    intents: NLU.IntentDefinition[],
+    entities: NLU.EntityDefinition[],
+    version: NLUVersionInfo,
+    lang: string
+  ): string {
+    const { nluVersion, langServerInfo } = version
+
+    const singleLangIntents = intents.map(i => ({ ...i, utterances: i.utterances[lang] }))
+
+    return crypto
+      .createHash('md5')
+      .update(JSON.stringify({ singleLangIntents, entities, nluVersion, langServerInfo }))
+      .digest('hex')
   }
 
   async train(
@@ -126,7 +142,7 @@ export default class Engine implements NLUEngine {
       ctxToTrain
     }
 
-    const hash = computeModelHash(intentDefs, entityDefs, Engine._version, languageCode)
+    const hash = this.computeModelHash(intentDefs, entityDefs, Engine._version, languageCode)
     const model = await this._trainAndMakeModel(input, hash, reportTrainingProgress)
     if (!model) {
       return
